+++ app/models/design/design.update.server.ts
@@ -19,3 +19,57 @@ export const updateDesignVisible = ({
 		data: { visible },
 	})
 }
+
+export const connectPrevAndNextDesigns = ({
+	prevId,
+	nextId,
+}: {
+	prevId: IDesign['id']
+	nextId: IDesign['id']
+}) => {
+	const connectNextToPrev = prisma.design.update({
+		where: { id: prevId },
+		data: { nextId },
+	})
+	const connectPrevToNext = prisma.design.update({
+		where: { id: nextId },
+		data: { prevId },
+	})
+	return [connectNextToPrev, connectPrevToNext]
+}
+
+export const updateDesignToHead = ({ id }: { id: IDesign['id'] }) => {
+	return prisma.design.update({
+		where: { id },
+		data: { prevId: null },
+	})
+}
+
+export const updateDesignToTail = ({ id }: { id: IDesign['id'] }) => {
+	return prisma.design.update({
+		where: { id },
+		data: { nextId: null },
+	})
+}
+
+export const updateDesignRemoveNodes = ({ id }: { id: IDesign['id'] }) => {
+	return prisma.design.update({
+		where: { id },
+		data: { prevId: null, nextId: null },
+	})
+}
+
+export const updateDesignNodes = ({
+	id,
+	nextId,
+	prevId,
+}: {
+	id: string
+	nextId: string | null
+	prevId: string | null
+}) => {
+	return prisma.design.update({
+		where: { id },
+		data: { prevId, nextId },
+	})
+}
