+++ app/services/design/move-up.service.ts
@@ -2,11 +2,13 @@ import { type User } from '@prisma/client'
 import { getDesign } from '#app/models/design/design.get.server'
 import {
 	type IDesign,
-	updateDesignRemoveNodes,
-	updateDesignNodes,
 	type IDesignEntityId,
 } from '#app/models/design/design.server'
-import { type IDesignUpdatedResponse } from '#app/models/design/design.update.server'
+import {
+	updateDesignRemoveNodes,
+	type IDesignUpdatedResponse,
+	updateDesignNodes,
+} from '#app/models/design/design.update.server'
 import { type designTypeEnum } from '#app/schema/design'
 import { type IUpdateSelectedDesignStrategy } from '#app/strategies/design/update-selected.strategy'
 import { prisma } from '#app/utils/db.server'
